requirements.txt
bitsandbytes==0.41.3.post2
dataloader==2.0
decord==0.6.0
deepspeed==0.9.5
editdistance==0.8.1
einops==0.8.0
fastapi==0.111.0
fire==0.5.0
flash_attn==2.4.2
gradio==4.31.5
huggingface_hub==0.22.2
imageio==2.13.5
jsonargparse==4.26.1
lightning==2.2.0rc0
lightning_utilities==0.9.0
matplotlib==3.5.1
nlg_metricverse==0.9.9
numpy==1.23.0
numpy==1.20.0
openai==0.28.0
opencv_python==4.5.5.64
packaging==21.3
pandas==1.3.4
peft==0.8.2
Pillow==9.0.0
pycocoevalcap==1.2
pyrender==0.1.45
pytorchvideo==0.1.5
quantize==0.0.4
ray==2.23.0
Requests==2.32.2
scipy==1.13.1
sentence_transformers==2.2.2
sentencepiece==0.1.99
Shapely==2.0.4
shortuuid==1.0.13
smplx==0.1.26
tokenizers==0.13.3
torch==2.0.0
torch_xla==2.3.0
torchvision==0.15.1+cu117
tqdm==4.66.1
transformers==4.28.1
trimesh==3.22.1
typing_extensions==4.12.0
uvicorn==0.30.0
visualize==0.5.1
xformers==0.0.22
